Refactor work for put endpoint
